Add format styling to the hero
The header is grouped pushed to the top to create isolation from the
main body of the hero. The hero main, secondary, and button are grouped
to draw focus to the sign up button. The image is positioned to add
character to the landing as this is first section a user will see.